Blood_supply/auth_service/requirements.txt
Django==5.0.7
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
django-environ==0.11.2
mysqlclient==2.2.4
PyJWT==2.8.0
python-dateutil==2.8.2
